minor fixes to the file handler and added tests

+++ lib/handlers/file.rb
@@ -49,8 +49,8 @@ module SSC
       require_appliance_id
       allow_remote_option
       def remove(file_name)
-        @file_list= FileListFile.new
-        file_id= @file_list.is_uploaded?(file_name)
+        file_list= FileListFile.new
+        file_id= file_list.is_uploaded?(file_name)
         if options.remote? && file_id
           begin
             StudioApi::File.find(file_id).destroy
@@ -61,8 +61,8 @@ module SSC
         elsif options.remote? && !file_id
           raise Thor::Error, "File '#{file_name}' not found"
         else
-          @file_list.push('remove', {file_name => nil})
-          @file_list.save
+          file_list.push('remove', {file_name => nil})
+          file_list.save
           say "File '#{file_name}' marked for removal"
         end
       end
@@ -70,51 +70,89 @@ module SSC
       desc 'file show FILE_NAME', 'show the contents of the file'
       require_appliance_id
       allow_remote_option
+      method_option :file_id, :type => :string
       def show(file_name)
         if options.remote?
-          id= find_file_id(file_name)
-          response= StudioApi::File.find(id)
-          say response.content
+          begin
+            require_appliance_directory do |appliance, files|
+              id= files[:file_list].is_uploaded?(file_name)
+              if id
+                response= StudioApi::File.find(id)
+                say response.content
+              else
+                say "File hasn't been uploaded.\nLocal Copy:", :red
+                say ApplianceDirectory.show_file(File.join('files',file_name))
+              end
+            end
+          rescue ApplianceDirectoryError
+            if options.file_id
+              say StudioApi::File.find(options.file_id).content
+            else
+              files= StudioApi::File.find(:all)
+              files= files.select {|f| f.filename == file_name}
+              raise Thor::Error, "File not found or ambiguous file name " unless files.length == 1
+              say files[0].content
+            end
+          end
         else
-          say show_file(file_name)
+          say ApplianceDirectory.show_file(File.join('files', file_name))
         end
       end
 
       desc 'file diff FILE_NAME', 'show the diff of the remote file and the local one'
       require_appliance_id
       def diff(file_name)
-        begin
-          id= find_file_id(file_name)
-          file_content= StudioApi::File.find(id).content
-        rescue
-          say "unable to connect or not in appliance directory", :red
-        end
-
-        begin
+        require_appliance_directory do |appliance, files|
+          id= files[:file_list].is_uploaded?(file_name)
+          raise Thor::Error, "File hasn't been uploaded" unless id
+          response= StudioApi::File.find(id)
+          remote_content= response.content
+          local_file= File.join(Dir.pwd, 'files', file_name)
           tempfile=Tempfile.new('ssc_file') 
-          tempfile.write(file_content)
-          say find_diff(tempfile.path, full_local_file_path(file_name))
+          tempfile.write(remote_content)
+          say find_diff(tempfile.path, local_file)
           tempfile.close; tempfile.unlink
-        rescue Errno::ENOENT
-          say "diff not installed", :red
         end
+      rescue ApplianceDirectoryError
+        raise Thor::Error, "diff can only be performed in the appliance directory"
       end
 
       desc 'file list', 'show all overlay files'
       require_appliance_id
       allow_remote_option
       def list
-        require_appliance do |appliance|
-          out= if options.remote? || file_list_empty?
+        require_appliance_directory do |appliance, files|
+          file_list= files[:file_list]
+          out= if options.remote? || file_list.empty_list?
             response= StudioApi::File.find(:all, :params => {:appliance_id => appliance.id})
-            response.collect do |file|
-              {file.filename => {"id" => file.id, "path" => file.path}}
+            response= response.collect do |file|
+              item= {file.filename => {"id" => file.id, "path" => file.path}}
+              file_list.push('list', item)
             end
+            file_list.save
+            response
           else
-            list_local_files
+            file_list["list"]
           end
           say out.to_yaml
         end
+      rescue ApplianceDirectoryError
+        require_appliance do |appliance|
+          print_table StudioApi::File.find(:all, :params => {:appliance_id => appliance.id}).collect do |file|
+            [file.id, File.join(file.path, file.filename)]
+          end
+        end
+      end
+
+      private
+
+      def get_file_list(appliance_id)
+      end
+
+      def find_diff(file1, file2)
+        `diff #{file1} #{file2}`
+      rescue Errno::ENOENT
+        raise Thor::Error, "'diff' not installed"
       end
     end
   end
